[Refactor]: CommunityPost.java 수정 및 패키지 이동
community domain 패키지로 이동
validation 필요한 필드 추가

Related to: #559
