+++ src/main/java/org/sopt/makers/internal/dto/community/CommunityPostMemberVo.java
@@ -1,7 +1,6 @@
 package org.sopt.makers.internal.dto.community;
 
 import com.querydsl.core.annotations.QueryProjection;
-import org.sopt.makers.internal.domain.community.CommunityPost;
 
 public record CommunityPostMemberVo(
         MemberVo member,
