+++ src/main/java/org/sopt/makers/internal/dto/community/PostCategoryDao.java
@@ -2,7 +2,7 @@ package org.sopt.makers.internal.dto.community;
 
 import com.querydsl.core.annotations.QueryProjection;
 import org.sopt.makers.internal.community.domain.category.Category;
-import org.sopt.makers.internal.domain.community.CommunityPost;
+import org.sopt.makers.internal.community.domain.CommunityPost;
 
 public record PostCategoryDao(
 		CommunityPost post,
