+++ src/main/java/org/sopt/makers/internal/mapper/CommunityResponseMapper.java
@@ -6,7 +6,7 @@ import org.sopt.makers.internal.domain.MemberCareer;
 import org.sopt.makers.internal.domain.community.AnonymousCommentProfile;
 import org.sopt.makers.internal.domain.community.AnonymousPostProfile;
 import org.sopt.makers.internal.community.domain.category.Category;
-import org.sopt.makers.internal.domain.community.CommunityPost;
+import org.sopt.makers.internal.community.domain.CommunityPost;
 import org.sopt.makers.internal.dto.community.*;
 import org.springframework.stereotype.Component;
