+++ src/main/java/org/sopt/makers/internal/repository/PostRepository.java
@@ -1,6 +1,6 @@
 package org.sopt.makers.internal.repository;
 
-import org.sopt.makers.internal.domain.community.CommunityPost;
+import org.sopt.makers.internal.community.domain.CommunityPost;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
