+++ src/main/java/org/sopt/makers/internal/repository/community/AnonymousPostProfileRepository.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 
 import org.sopt.makers.internal.domain.Member;
 import org.sopt.makers.internal.domain.community.AnonymousPostProfile;
-import org.sopt.makers.internal.domain.community.CommunityPost;
+import org.sopt.makers.internal.community.domain.CommunityPost;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AnonymousPostProfileRepository extends JpaRepository<AnonymousPostProfile, Long> {
