+++ src/main/java/org/sopt/makers/internal/repository/community/CommunityQueryRepository.java
@@ -5,6 +5,7 @@ import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import lombok.val;
+import org.sopt.makers.internal.community.domain.CommunityPost;
 import org.sopt.makers.internal.community.domain.category.QCategory;
 import org.sopt.makers.internal.domain.*;
 import org.sopt.makers.internal.domain.community.*;
